pyproject.toml
[build-system]
requires = ["setuptools", "setuptools-scm"]
build-backend = "setuptools.build_meta"

[project]
name = "arcueid"
requires-python = ">=3.10"
dynamic = ["version"]
readme = "README.md"
description = "A simple discord bot for finding groups of friends to talk to!"
dependencies = [
    "discord.py==2.0.0",
    "click==8.1.3",
    "pytz"
]

[tool.setuptools.dynamic]
version = { attr = "arcueid.VERSION" }

[project.scripts]
arcueid = "arcueid.__main__:launch"